refactor(agile-dashboard-view): architectural improvements with additional highly-cohesive submodules
