+++ src/features/agile-dashboard-view/domain/view-state.ts
@@ -1,0 +1,7 @@
+export type SelectedView = "projects" | "completed";
+
+export interface DashboardState {
+	selectedView: SelectedView;
+	activeOnly: boolean;
+	selectedAlias: string | null;
+}
